do user check at top level
